Reworked logic for removing invisible border cubes
My previous optimization was a chihuahua dog shit, and I'm fucking fed up that I throw up my previous design which was correct which was the design I implement right now. However I'm happy that it's working now and there is still a lot of work to remove redundant and death code to make MapOptimization script cleaner.